The "Balance" report was moved to separate file.
#63
